+++ CHANGELOG.md
@@ -6,6 +6,7 @@ HEAD
 
 * Remove unnecessary class instance
 * Remove IWebConfig dependency and replace parseWebConfig with setConfig method
+* Fix setup definition
 
 v0.0.4 (2023-05-08)
 ------